Fix halt and resume event bug by adding a close blocking modal event. (PR #6549)
# Description

I added a `close-blocking-modal` event to make the implementation less confusing. Previously, the `resume-event` or `halt-event` were fired a second time to trigger the closure of the modal. With a previous refactor, I removed the "redundant" event calls, thinking it was a mistake, with the unwanted result that the modal would stay open.

This only affects iso9.

closes:
- #6548

https://github.com/user-attachments/assets/b0588a4b-81bf-4152-b371-4e16f505dcd7