Add a page to view the details of a discovered resource. (Issue #6562, PR #6566)
# Description

Add a page to display the details of a discovered resource, matching the behaviour on the Resource page.
The links forwarding to the Managed/discovery resource are now in a kebab menu at the end of the row. When no links ara available, it is disabled to keep the space it uses in the row, preventing the "show details" button to not be aligned properly in rows that don't have links.

**Note:**
> - Sorting is temporarily disabled, until the API supports filtering by agent/value/type.
> - E2E tests cannot be implemented, the local-setup doesn't have discovered resources.
> - The name of the branch is not accurate, I was innitially supposed to start with the filtering, but had to switch to working on the page instead.

closes:
- #6562

https://github.com/user-attachments/assets/e944cc56-09c0-48c1-adbe-d5ee32435481